API/requirements.txt
Flask==2.1.1
flask-cors==3.0.10
python-dotenv==0.20.0
pymongo[srv]==4.1.1